The frontend is done !

+++ src/components/WorkoutDetails.js
@@ -1,12 +1,21 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { formatDistanceToNow } from "date-fns";
 
 const WorkoutDetails = ({ workout }) => {
+	const { user } = useAuthContext();
 	const { dispatch } = useWorkoutsContext();
 
 	const handleClick = async () => {
+		if (!user) {
+			return;
+		}
+
 		const response = await fetch("/api/workouts/" + workout._id, {
 			method: "DELETE",
+			headers: {
+				Authorization: `Bearer ${user.token}`,
+			},
 		});
 		const json = await response.json();
